Add files via upload
This Sustained Attention to Response Task works with Psychtoolbox installed on MATLAB. 
It was created under the specifications by Robertson et al. (1996). Please read the article included. 
the script start by executing the command "sart". It will ask to indicate the experimenter initials and participant number at the beginning. 
It is possible to automatically create a .xls with the results from participants tested by executing the command "printfile".